+++ internal/controller/keycloak.go
@@ -95,6 +95,9 @@ func GetKeycloakClient(ctx context.Context, server GoCloakServer, keycloakClient
 			return nil, err
 		}
 		client.Secret = secret.Value
+	} else {
+		emptyStr := ""
+		client.Secret = &emptyStr
 	}
 	return client, nil
 }