+++ test/e2e/e2e_test.go
@@ -47,6 +47,8 @@ const metricsRoleBindingName = "keycloak-cr-operator-metrics-binding"
 
 const keycloakClientManifest = "config/samples/keycloak_v1alpha1_keycloakclient.yaml"
 const keycloakClientManifestWithSecret = "config/samples/keycloak_v1alpha1_keycloakclient_with_secret.yaml"
+const keycloakClientManifestPublic = "config/samples/keycloak_v1alpha1_keycloakclient_public.yaml"
+const keycloakClientManifestHeadlamp = "config/samples/keycloak_v1alpha1_keycloakclient_headlamp.yaml"
 
 const deploymentYaml = "test/e2e/deployment.yaml"